add download/upload token API and client support.
API is added for getting token for operations like download/upload.
The client is updated to use the token API to get tokens for download
operations. The token is cached execution time during downloads
according to RSE and it tokens. Minimum lifetime is given before we
call the token API to refresh the token. And the retry happens with
the new token.